Fixed up location of base class. Addressed warnings of unreferenced attribute values.
